Clean up AP - refresh token for suth, catalog param
Allow user to specify the catalog number to restore into. The restore will fail if the catalog does not already exist - you cannot set your own catalog ID. Also save a refresh token for authentication instead of a (fast-expiring) access token. In the future, this won't be reliant on a preset token.