Fixed bug in logging
Logging in actions.py wasn't properly being logged to file. This was
due to multiple loggers being in use -- the main log used the log
from multiprocessing and actions.py used its own.

In essence, the multiprocessing log appends $process to the log output,
so I setup logging to use the cfde_ap logger instead and added $process
to it. This way, the entire cfde_ap package logs properly and also
includes different process_ids in the logfiles